+++ datadog/check.go
@@ -3,6 +3,6 @@ package datadog
 // contains primitives for working with objects that datadog expects
 
 type CheckConf struct {
-	InitConfig []interface{} `yaml:"init_config"`
-	Instances  []interface{} `yaml:"instances"`
+	InitConfig map[string]interface{} `yaml:"init_config"`
+	Instances  []interface{}          `yaml:"instances"`
 }