+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 i## Changelog
 
+### [v0.2.6](https://github.com/dansteen/consuldog/releases/tag/v0.2.6) - 01 Aug 2017 
+Fixed issues with datadog process detection and signaling
+
+#### Breaking Changes
+
+#### Bug Fixes
+
+#### Improvements
+
+### [v0.2.4](https://github.com/dansteen/consuldog/releases/tag/v0.2.4) - 31 Jul 2017
+Fixed issues with yaml template processing
+
+#### Breaking Changes
+
+#### Bug Fixes
+
+#### Improvements
+
 ### [v0.2.1](https://github.com/dansteen/consuldog/releases/tag/v0.2.1) - 28 Jul 2017
 Initial release